requirements.txt
langchain==1.2.10
langgraph==1.0.10
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
chromadb==1.5.2
sentence-transformers==5.2.3
pymupdf==1.27.1
requests==2.32.5
python-dotenv==1.2.2
